requirements/development.txt
-i https://pypi.python.org/simple
-r base.txt
django-debug-toolbar==2.0
django-storages[google]>=1.9.1
pre-commit==2.5.1
pyyaml==5.3.1
sqlparse==0.3.0
werkzeug==0.16.0
